Rewrite Config class to use AboutData and platformdirs
Instead of hardcoding paths use platformdirs for configuration-,
pictures- and cache-directory.

Also instead of creating a global instance, initialize the class
in the constructor, including the creation of all directories we use.

Also it's not required to create the config file if it doesn't exist,
configparser will simply not an emtpy configuration object.

Change default value of wallpaper to be a string, the
config parser later uses a string, if the value is a list as it
was said it doesn't work a list doesn't have a .split() method.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>