fix: config won't load if they're not backends detected
Even when setting `backend = none` in your config, the program refuses
to start since it assumes it has detected at least 1 backend, which with
the "none" backend would make it 2. "none" isn't really a backend though, and
the code on `__init__` fails if no backends were detected before it has the
chance to load the config.